requirements.txt
-e .
pytest
aws-cdk.core
aws-cdk.aws-ec2
aws-cdk.aws-rds
